Feat: client-side routing
Added some components to establish the structure
of the page. With these components, there is a
home page, a register page, a log in page and an
error page. Each page has its own style rules.
There's also styles that support lower device
sizes.

Now, to make the routing, it uses the react-dom
library. Added a routes file that handles this.
This file is being used by the main.jsx file.It
also renders the error page when a user tries
to visit a page that doesn't exists.

The App.css file contains some colors used in
every page. Meanwhile, in the index.css it's used
the modern-normalize stylesheet to fix some
inconsistencies between browser styles.

Lastly, also added propTypes library to define
a propType in the Form component. Another minor
addition is a useRef to autofocus an input field
in both register and log in pages when rendered.