Add: validation icons
The current page lacks validation feedback. A user
doesn't know when his password or email is correct
or wrong. These new icons will helps solve this
issue.

Another thing to keep in mind is to avoid some
repetition when rendering these new icons. So they
are conditionally rendered in a FormChild
component, only when it contains an input as part
of the children props; hence the new "isInput"
prop.