Feat: validation icons animations
Before this commit, the icons didn't give actual
feedback to the user when entering his data. With
some CSS animations and a conditional located in
Input.jsx, icons now change as the user types the
fields.

Now it also changes the input's background color
accordingly to the validity of the data. It also
changes from invalid to valid or the other way.

Finally, there are some minor changes in the
PasswordInput and Input component to allow a prop
that is used to conditionally assign an "invalid"
class name to their respective input element and
apply styles.