Add: PasswordStrengthMeter component
Having only basic password validation is not
enough to meet certain security requirements. This
new component utilizes the "zxcvbn" library. It
also renders a progress bar that changes color
based on the password's strength.

Another change is to pass a prop through
PasswordInput and Input components to
conditionally render an "invalid" class name in
the input element, just like the previous commit.