Feat: user data validation at register account
Without data validation, users can enter anything
when trying to create an account. These methods
ask the users to enter a valid email direction.
And a valid password.

For this application, a valid password must be
at least 8 characters long, counting each unicode
character as one. And with the use of the "zxcvbn"
library, password's score must be higher than 2.

Another change is modifying the Form component to
accept a handleSubmit method and adding the
"noValidate" attribute. This last one is to
prevent HTML 5 validation applying the wrong
styles to inputs.