Add: context to several components
Login and Register components have some
inconsistencies. One of the ugliest is that
input icons are not placed correctly in one of the
form components. This happens because they share
certain values used to position said elements.

To solve this issue, now the Login and Register
components have a React Context variable and a
provider to pass a value. With these change, child
components will prepend the form component name to
form new CSS classes. Lastly, in CSS files, each
one has its own values, ensuring that icons are
placed correctly.