Add: PasswordStrengthMeter to PasswordInput
This component is added now because for this page
password strength meter is meant to guide the user
when registering an account. So it should not be
displayed on Login component.

Thanks to the addition of Context in last commit,
now is the time to add this component. Another
minor change is the adjustment of this component.
It will stay just below the password input.