Add: user data validation in backend
Despite having user data validation in frontend,
the app is not secure enough. Some users will
bypass that validation.

With this user data validation, it will halt some
malicious attempts. Like in frontend, it will ask
the users to enter a valid email and a "valid"
password. It also uses the Zxcvbn library to
validate passwords.