Feat: handle user requests at backend
Since this project utilizes a React app as
frontend, Ruby app is set as API mode. Thus, it
is not necessary to render views inside
controllers. Just send responses to frontend as
JSON.

It also adds logic to handle account creation and
deletion. It also manages a session creation and
deletion (e.g. user logs in / out).