Add: Axios requests
Before this commit, users can enter data and
backend is able to validate it. But Frontend needs
to let backend know the users intentions.

With Axios library, requests are easier to read
than using the built in fetch API. This
application now can make POST and DELETE requests.
These are used to create an account, log in and
log out.

In a future commit, frontend will redirect the
user when it creates or logs in / out.