Feat: page navigation
Previously, requests were added. It lacked page
redirection when the user created an account, log
in or log out.

The new useGoToPage hook eases this "new" feature.
It uses the React Router Dom useNavigate hook to
navigate between pages. It also accepts the
replace attribute. This one is used to prevent
user from returning to one of the form pages after
being redirected.