Add: user email persistence and rewrite routes
After a user logs in, if he refreshed or closes
and reopens, it will get "log out". This happens
because data is not stored by default.

To achieve this change, routes file was modified.
It also uses the Outlet and OutletContext from
React Router Dom to pass the user email from the
App component to its children, now the form
components at the moment.

There is another change in Home component. It
conditionally renders button when it detects a
user has logged in. Also, in the form components
when it detects a user it redirects them to
another page (to dashboard, but it has not been
implemented yet).