Refactor: user requests
Users requests were scattered through each form
component and Home component. Before adding more
features, it is a good moment to rewrite this
part.

There is a new hook that contains these requests.
Some duplicated code from request methods
was now reduced thanks to this new hook. It also
receives the setErrorMessage and setUserEmail
since they are used in every request. The formData
is optional because it is not needed to make a
"log out" request.