Refactor: user authentication
Previous user authentication was really simple.
Also, it wasn't secure. This new strategy utilizes
tokens to authenticate users.

In backend, it utilizes the devise_jwt gem to
apply this strategy; since this project already
uses devise gem for authentication. Between the
token revocation strategies available, it uses one
called "deny list". It needs a new table to work.

Jsonapi-serializer gem was installed to serialize
some data from successful responses and send it to
frontend.

Some configuration changes were made across
several files in backend in order to get this
feature working. Also, there is a new file that
prevents an server error when the user registers.

In frontend there also are many changes that allow
to send the data correctly and manage user tokens
from server responses (storing or deleting).

In many apps, when a jwt token expires, it should
use a refresh token to make a request to backend
in order to fetch a new jwt token. According to
devise_jwt's documentation, it does not implement
the strategy mentioned before. The creator
recommends to install another gem instead.

Another alternative is to log out user when jwt
token expires and redirect to log in page. These
few changes in useUserAuthentication hook
implement this idea. There is an empty catch block
in delete request. It prevents displaying a 401
error in browser's console.

One thing that should be added is a notification
for users when jwt tokens expire before logging
them out.