Fix: loading error in dev server
According to Vite's documentation, when a
React file exports anything that is not a
component, it breaks the HMR feature. Login and
Register files had a context const that caused
this issue.

Now there is a new contexts folder, with new files
also. And the Login, Register and other files now
import said contexts from those new files.