Feature: Dashboard component in frontend
Previously, when user registers or logs in, it
redirects to dashboard page, but said page was
empty. This components is a simple addition to
this project.

This component contains some simple elements: a
navbar with the user email and a button to log
out. Also there is a vertical navbar that shows
registered users. And lastly, another div that
will show a chat selected from the user in a
future commit.

A minor problem at the moment is that requests
calls in frontend were made in only one file. That
happens because those calls are related to
authenticate users. With this new Dashboard, there
is a need to make more requests to backend.

Thanks to Axios instances it is possible to enter
a specific configuration that can be used in
several files and separate responsibilities. Also,
it is stored in its own file. And this means less
code repetition.

With this change, useUserAuthentication hook now
makes use of this new instance, as well this new
component.

Without a list of users, they can't communicate.
Now there is a new route in backend that returns
all the registered users emails so frontend can
fetch and render said list.

At the moment of writing this commit, user cannot
send messages. Another thing to change later is
that Dashboard page's styles are simple.