Feature: add messaging
Until now, users couldn't communicate between
themselves. Now it's possible to do that.

First, there is a new Message model in backend,
with some other configurations like associations,
routes, controllers, etc. that handles these
messages. A note about this model is that it only
supports conversations between two users.

Another important addition in backend is a scope
that returns messages between two users with their
emails attached in each message. This is useful
when rendering messages in frontend.

Speaking of frontend, there is now a registered
users list in Dashboard component. With this users
can select who to talk to. It has styles when each
user is hovered to let them know they can be
clicked. And each time a user is clicked it
fetches messages between them, if there are.

Another relevant change in frontend is scroll bars
for both users list and messages. This is done in
case there are many of either. So page does not
grows size and keeps looking "identical" as when
a user enters.

Now, the last change, is a form. With these,
combined with changes above, users can communicate
with others. It only appears when a user clicks
another's email. And always stay at the bottom
of the screen.

Lastly, there is a related known issue at the
moment of writing this commit: when a user's JWT
token expires, they can still write messages to
other users.